+++ ext/libarchive-ruby-swig/reader.h
@@ -44,18 +44,18 @@ class Reader
 %feature("autodoc", "Use <code>Archive::read_open_filename</code> instead") read_open_filename;
 %newobject read_open_filename(const char *filename, const char *cmd = 0, bool raw = false);
 %feature("autodoc", "Use <code>Archive::read_open_memory</code> instead") read_open_memory;
-%newobject read_open_memory(const char *string, int length, const char *cmd = 0, bool raw = false);
+%newobject read_open_memory(const char *string, size_t length, const char *cmd = 0, bool raw = false);
 %feature("autodoc", "Returns the next Entry meta data object in the Archive") next_header;
 %newobject next_header();
 #endif
 
         static Reader *read_open_filename(const char *filename,
             const char *cmd = 0, bool raw = false);
-        static Reader *read_open_memory(const char *string, int length,
+        static Reader *read_open_memory(const char *string, size_t length,
             const char *cmd = 0, bool raw = false);
 
         Entry *next_header();
-        VALUE read_data_helper(int len);
+        VALUE read_data_helper(size_t len);
 
 #ifdef SWIG
 %exception;
@@ -64,7 +64,7 @@ class Reader
     private:
         struct archive *_ar;
         char *_buf;
-        int _buf_size;
+        size_t _buf_size;
         char *_archive_content;
 };
